// frontend/src/app/interfaces/IProduto.interface.ts
export interface IProduto {
    id: number;
    nome: string;
    descricao?: string;
    precoVenda: number;
    imagemUrl ?: string;
    observacao?: string;
    unidadeMedida?: string;
    extra1?: string;
    extra2?: string;
    extra3?: string;
    extra4?: string;
    extra5?: string;
}